requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic~=2.12.5
pillow==10.4.0
lpips==0.1.4
torch~=2.10.0
torchvision~=0.25.0
numpy~=2.4.1